PDF settings, PDF actions, base64 generator

+++ src/main/java/com/tableManager/tableManager/controllers/PdfController.java
@@ -3,13 +3,21 @@ package com.tableManager.tableManager.controllers;
 
 import com.tableManager.tableManager.model.PdfSettings;
 import com.tableManager.tableManager.model.User;
-import com.tableManager.tableManager.service.PdfService;
-import com.tableManager.tableManager.service.UserService;
 import com.tableManager.tableManager.service.impl.PdfServiceImpl;
 import com.tableManager.tableManager.service.impl.UserServiceImpl;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.core.io.Resource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Base64;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api/pdf")
@@ -18,18 +26,63 @@ public class PdfController {
 
     private final PdfServiceImpl pdfService;
     private final UserServiceImpl userService;
+    private static final String UPLOAD_DIR = "src/main/resources/logos/";
 
     public PdfController(PdfServiceImpl pdfService, UserServiceImpl userService) {
         this.pdfService = pdfService;
         this.userService = userService;
     }
 
-    @GetMapping("/getPDFSettings/{userId}")
-    public ResponseEntity<PdfSettings> getPdfSettings(@PathVariable Long userId) {
+    @GetMapping("/base64")
+    public ResponseEntity<String> getImage() throws IOException {
+        Long currentUserId = userService.getCurrentUserId();
+        PdfSettings pdfSettings = pdfService.getPdfByUserId(currentUserId);
+
+        if (pdfSettings == null || pdfSettings.getLogoURL() == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        String logoURL = pdfSettings.getLogoURL();
+        Resource resource = new ClassPathResource(logoURL);
+
+        if (!resource.exists()) {
+            return ResponseEntity.notFound().build();
+        }
+
+        byte[] imageBytes = Files.readAllBytes(resource.getFile().toPath());
+        String base64 = Base64.getEncoder().encodeToString(imageBytes);
+
+        return ResponseEntity.ok(base64);
+    }
+
+    @PostMapping("/upload")
+    public ResponseEntity<Object> upload(@RequestParam("file") MultipartFile file) {
+        try {
+            if (file.isEmpty()) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(Map.of("error", "File is empty"));
+            }
+            String contentType = file.getContentType();
+            if (contentType == null || (!contentType.equals("image/png") && !contentType.equals("image/jpeg"))) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid file type. Only PNG and JPG files are allowed.");
+            }
+
+            String filename = file.getOriginalFilename();
+            Path path = Paths.get(UPLOAD_DIR + filename);
+            Files.createDirectories(path.getParent());
+            Files.write(path, file.getBytes());
+            String fileUrl = "src/logos/" + filename;
+            return ResponseEntity.ok(Map.of("fileUrl", fileUrl));
+        } catch (IOException e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Map.of("error", "Failed to upload file"));
+        }
+    }
+
+    @GetMapping("/getPdfSettings")
+    public ResponseEntity<PdfSettings> getPdfSettings() {
         Long currentUserId = userService.getCurrentUserId();
         User user = userService.findbyId(currentUserId);
 
-        if (currentUserId != null && currentUserId.equals(userId) && user.isEnabled()) {
+        if (currentUserId != null && user.isEnabled()) {
             PdfSettings pdfSettings = pdfService.getPdfByUserId(currentUserId);
             return ResponseEntity.ok(pdfSettings);
         } else{
@@ -37,7 +90,7 @@ public class PdfController {
         }
     }
 
-    @PostMapping("/postPDFSettings")
+    @PostMapping("/addPdfSettings")
     public ResponseEntity<PdfSettings> postPdfSettings(@RequestBody PdfSettings pdfSettings) {
         Long currentUserId = userService.getCurrentUserId();
         User user = userService.findbyId(currentUserId);
@@ -50,7 +103,7 @@ public class PdfController {
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
-    @PutMapping("/updatePDFSettings")
+    @PutMapping("/updatePdfSettings")
     public ResponseEntity<PdfSettings> updatePdfSettings(@RequestBody PdfSettings pdfSettings) {
         Long currentUserId = userService.getCurrentUserId();
         User user = userService.findbyId(currentUserId);
@@ -63,12 +116,25 @@ public class PdfController {
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
-    @DeleteMapping("/deletePDFSettings")
+    @DeleteMapping("/deletePdfSettings")
     public ResponseEntity<?> deletePdfSettings(){
         Long currentUserId = userService.getCurrentUserId();
         User user = userService.findbyId(currentUserId);
         if (currentUserId != null && user.isEnabled()) {
-            pdfService.deletePDF(currentUserId);
+
+            String filename = pdfService.getFilename(currentUserId);
+            Path path = Path.of(filename);
+
+            try{
+                Files.delete(path);
+            }catch (IOException e){
+                e.printStackTrace();
+            }
+
+            if(!Files.exists(path)) {
+                pdfService.deletePDF(currentUserId);
+            }
+
             return new ResponseEntity<>(HttpStatus.OK);
         }else{
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
